refactor(main): extract terrain, CRS, and inspector-card units from god-module
Reduce the src/main.ts god-module by moving three cohesive clusters into
focused modules under src/app/, with zero behavior change:

- terrainAnalysisRunner: the async terrain-analysis orchestration, the A-1
  stale-result token guard, the fingerprint core cache, and the worker
  offload. The runner owns its run-state (token, AbortController,
  cache-clear fn) and exposes run() + abortAndClearCache().
- crsCoordinator: CRS resolution, per-scan refresh (static + streaming),
  and user override handling. Owns only the per-scan override-store key;
  CrsService remains the source of truth.
- inspectorCardRefreshers: load-time Provenance + Dataset Intelligence
  card refreshers.

Stateful collaborators (lazy viewer, activeId, viewerReady) are passed as
getters so no top-level viewer.* dereference is introduced; the deferred
viewerLoaded discipline is preserved. main.ts: 4079 -> 3696 LOC.
